requirements.txt
flask
gunicorn
timm
pillow
numpy
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
